upd logic
