+++ packages/espressocash_app/lib/features/ramp/widgets/ramp_buttons.dart
@@ -1,7 +1,6 @@
 import 'dart:async';
 
 import 'package:dfunc/dfunc.dart';
-import 'package:fast_immutable_collections/fast_immutable_collections.dart';
 import 'package:flutter/material.dart';
 
 import '../../../../../l10n/l10n.dart';
@@ -12,7 +11,6 @@ import '../../../ui/icon_button.dart';
 import '../../accounts/models/account.dart';
 import '../../analytics/analytics_manager.dart';
 import '../../country_picker/models/country.dart';
-import '../../feature_flags/services/feature_flags_manager.dart';
 import '../../profile/data/profile_repository.dart';
 import '../../ramp_partner/models/ramp_partner.dart';
 import '../models/profile_data.dart';
@@ -162,12 +160,9 @@ extension RampBuildContextExt on BuildContext {
     required ProfileData profile,
     required String address,
   }) {
-    final partners = _getOnRampPartners(profile.country.code);
-
     RampPartnerSelectScreen.push(
       this,
       type: RampType.onRamp,
-      partners: partners,
       onPartnerSelected: (RampPartner partner) {
         Navigator.pop(this);
         _launchOnRampPartner(partner, profile: profile, address: address);
@@ -179,12 +174,9 @@ extension RampBuildContextExt on BuildContext {
     required ProfileData profile,
     required String address,
   }) {
-    final partners = _getOffRampPartners(profile.country.code);
-
     RampPartnerSelectScreen.push(
       this,
       type: RampType.offRamp,
-      partners: partners,
       onPartnerSelected: (RampPartner partner) {
         Navigator.pop(this);
         _launchOffRampPartner(partner, profile: profile, address: address);
@@ -239,70 +231,3 @@ extension RampBuildContextExt on BuildContext {
         .rampOpened(partner: partner, rampType: RampType.offRamp.name);
   }
 }
-
-IList<RampPartner> _getOnRampPartners(String countryCode) {
-  final partners = <RampPartner>{};
-
-  if (_kadoCountries.contains(countryCode)) {
-    partners.add(RampPartner.kado);
-  }
-
-  if (_scalexCountries.contains(countryCode)) {
-    partners.add(RampPartner.scalex);
-  }
-
-  partners.add(RampPartner.rampNetwork);
-
-  if (_guardarianCountries.contains(countryCode)) {
-    partners.add(RampPartner.guardarian);
-  }
-
-  final isMoneygramEnabled =
-      sl<FeatureFlagsManager>().isMoneygramAccessEnabled();
-
-  if (isMoneygramEnabled && _moneygramOnRampCountries.contains(countryCode)) {
-    partners.add(RampPartner.moneygram);
-  }
-
-  return IList(partners);
-}
-
-IList<RampPartner> _getOffRampPartners(String countryCode) {
-  final partners = <RampPartner>{};
-
-  if (_coinflowCountries.contains(countryCode)) {
-    partners.add(RampPartner.coinflow);
-  }
-
-  if (_scalexCountries.contains(countryCode)) {
-    partners.add(RampPartner.scalex);
-  }
-
-  final isMoneygramEnabled =
-      sl<FeatureFlagsManager>().isMoneygramAccessEnabled();
-
-  if (isMoneygramEnabled && _moneygramOffRampCountries.contains(countryCode)) {
-    partners.add(RampPartner.moneygram);
-  }
-
-  return IList(partners);
-}
-
-const _kadoCountries = {'US'};
-
-const _guardarianCountries = {
-  'AT', 'BE', 'BG', 'HR', 'CY', 'CZ', 'DK', 'EE', 'FI', 'FR', 'DE', 'GR', //
-  'HU', 'IE', 'IT', 'LV', 'LT', 'LU', 'MT', 'NL', 'PL', 'PT', 'RO', 'SK',
-  'SI', 'ES', 'SE', 'IS', 'LI', 'NO', 'CH',
-};
-
-const _coinflowCountries = {
-  'AD', 'AT', 'BE', 'BG', 'HR', 'CY', 'CZ', 'DK', 'EE', 'FI', 'FR', 'DE', //
-  'GR', 'HU', 'IS', 'IE', 'IT', 'LV', 'LI', 'LT', 'LU', 'MT', 'MC', 'NL', 'NO',
-  'PL', 'PT', 'RO', 'SM', 'SK', 'SI', 'ES', 'SE', 'CH', 'US',
-};
-
-const _scalexCountries = {'NG'};
-
-const _moneygramOnRampCountries = {'US'};
-const _moneygramOffRampCountries = {'US', 'PT'};
